nodal linear and cubic
